PitchScope Player is a MP3 player which can also do realtime note detection
PitchScope Player is a type of music analysis software that is designed
to detect the most dominant pitch (a musical note) at any given moment
in time within a MP3 or WAV music file. Player is designed to work on
musical instrument solos, and is not really intended to detect the pitch
of a human voice. It can not detect multiple simultaneous notes, but
will resort to just finding the loudest pitch at any given moment in
time. The C++ source code can be compiled with Microsoft Visual Studio.
PitchScope Player was designed and built by James Paul Millard (dba
Creative Detectors of California).